Implementado controladores y servicios

+++ src/main/java/com/parcial/parcial02/service/MagicProviderService.java
@@ -3,7 +3,9 @@ package com.parcial.parcial02.service;
 import com.parcial.parcial02.dto.CreateProviderDto;
 import com.parcial.parcial02.dto.ProviderDto;
 import com.parcial.parcial02.dto.ProviderMapper;
+import com.parcial.parcial02.exception.BusinessRuleException;
 import com.parcial.parcial02.exception.EntityNotFoundException;
+import com.parcial.parcial02.repository.MagicArticleRepository;
 import com.parcial.parcial02.repository.MagicProviderRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -15,6 +17,7 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class MagicProviderService {
     private MagicProviderRepository providerRepository;
+    private MagicArticleRepository articleRepository;
 
     public ProviderDto createProvider(CreateProviderDto request) {
         var newProvider = ProviderMapper.toEntity(request);
@@ -39,12 +42,29 @@ public class MagicProviderService {
         return ProviderMapper.toDto(provider);
     }
 
+    public ProviderDto updateProvider(CreateProviderDto request, UUID id) {
+        var provider = providerRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Proveedor no existe"));
+
+        provider.setName(request.name());
+        provider.setType(request.type());
+
+        providerRepository.save(provider);
+
+        return ProviderMapper.toDto(provider);
+    }
+
     public void deleteProvider(UUID id) {
 
         var provider = providerRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Proveedor no existe"));
 
+        var articles = articleRepository.findByProviderId(id);
+
+        if(!articles.isEmpty()) {
+            throw new BusinessRuleException("No se puede eliminar el proveedor");
+        }
+
         providerRepository.delete(provider);
     }
-
 }
